fix(cactus-web): fix a couple of blur-related bugs
CACTUS-98

Calling `setTimeout` in a blur handler results in code being run after
the new element has focus, in all browsers I tested. I'd still prefer to
use `relatedTarget`, but we don't always get what we like.

I also changed the element that receives focus when you click the scroll
buttons, so that we don't have to kill blur events when the button
disappears; it also made the behavior more consistent across browsers.

Finally, I made it so the scroll state of a menu resets when it loses
focus.

I did see a weird issue in IE, but couldn't reproduce it with the
console open: when I opened a drop-down the length would fluctuate. I
think it had something to do with how opening the menu added a vertical
scroll bar to the page, in which case I'm not too worried because on a
normal-sized screen, with the menu bar at the top, it should almost
never cause that kind of overflow.