fix(cactus-web): fix scroll icons and action handlers
CACTUS-98

It turns out that <Enter> and <space> were behaving differently because
they trigger the button's `onClick` handler on different events: <Enter>
triggers on `keypress` and <space> triggers on `keyup`. As such, I was
unable to rely on the default behavior, so I disabled and replaced it.
To get it to work more smoothly, I split the submenu handler in two, one
for the button and one for the overall submenu wrapper.