fix(cactus-web): fix a couple of focus issues
CACTUS-98

Fixes the fact that sometimes when scrolling using the keyboard, the
focused element isn't visible. Unfortunately, this fix causes the scroll
handler to be run twice, so I still need to look into a way of fixing
that. It also changes the way the first element gets tab focus, from a
dynamic handler to a somewhat more static version.