refactor(cactus-web): decouple Sidebar code from Layout
In order to decouple Sidebar, I changed Layout from directly referencing
the Sidebar `styled-components` class, to referencing a generic CSS
class that could be assigned to any component. This will also make it
easier to cascade styles down the tree based on the current layout
position.

While I was at it, I did a little refactoring of the `Footer`.