fix(cactus-web): fix nested state update issue & expand test
CACTUS-346

The issue seemed to be that I was calling a state update (setFocus)
inside another state update callback (for setExpanded), and React
doesn't always call those callbacks immediately, so the second call
wasn't even picked up until the re-render had already started. (This
also happens in the browser, but doesn't cause a console warning for
some reason.) To avoid the need for a state callback while keeping the
callback references stable, I added a "box" ref, an immutable container
with mutable contents.