feat(cactus-web): change breadcrumb colors to pull from the theme and match the Style Guide
The colors were previously hard-coded, and we were not adding the proper hover and focus effects

BREAKING CHANGE: The colors of the Breadcrumb.Item links now pull from the theme.  In addition, the
color now changes when the links are hovered, and a custom border appears when the links are
focused.

CACTUS-329