fix(cactus-web): workaround for styled components bug (?) with nested styles
It turns out styled-components doesn't properly update dynamic styles of
the format `* &`. Static styles work fine, but if there's anything
dynamic about them it creates the same style multiple times and only
applies the most recently created (meaning if you toggle a style back
and forth, it will probably get stuck in one of the styles).

To work around this on the Layout component I moved the "dynamicness" up
to the layout wrapper where those classes are defined, so it treats the
parent as the dynamic component instead of the descendant.

I did not check for other occurrences of this bug.