fix(cactus-web): fix auto-scrolling behavior of MenuBar
CACTUS-459

The problem is basically a conflict between browsers' auto-scroll when
an element receives focus, and the scroll behavior implemented by the
hook. To get around this, we have to disable the hook scroll when focus
is the result of a click, and then do the scroll after the click. This
doesn't 100% fix the problem, though: if you click a submenu item to
expand it and the expanded part is completely off-screen, the
auto-scroll will still kick in (there's no cross-browser way to prevent
that behavior).

This seemed like a good time to refactor the scroll behavior used by
both MenuBar and TabList, so I also did that and refined the targeted
scrolling algorithm a bit to make it flow a little more naturally.