test(mock-ebpp): add integration tests for DateInput
CACTUS-476

Since I couldn't think of a way of working the datetime/time inputs into
the example app, I made a "hidden" page specifically for the integration
tests. With that done, I felt it made more sense to move the existing
DateInput tests to that file as well.

Finally, I also fixed an issue where IE (I think it's actually a
polyfill library) was adding an extra character to the month names.