fix(cactus-web): fix mac focus issue for `button` elements
CACTUS-485

Mac (for some reason) does not like focusing buttons on click, which led
to the list being focused and triggering the "focus on first menuitem"
code. The best solution is to swap out `button` for another element: in
this case I chose `span` for its similarity to `a`, which is probably
the most commonly used `as` prop for `MenuBar.Item`.