feat(cactus-web): clear selection when value="" is passed (Select & FileInput)
CACTUS-567

It's difficult to simulate adding a file in unit tests, so `FileInput`
is covered by a change in Storybook instead.

Select obviously has changes way beyond the apparent scope of the
ticket; but when I was trying to find the place I had to modify, I
discovered that the way value & options & extraOptions interacted didn't
make any sense (some stuff was happening before render, some during
render, and some after). So I moved it all to one place so it flows in
an easily traceable, linear fashion.