feat(cactus-web): add prop-forwarding to AccessibleField
CACTUS-597

This one's a little tricky. On the one hand, it's hard to say there's no
reason to forward props; on the other hand, some props would not work
very intuitively so for those it would be best to just avoid them: for
example, if I as a user passed `onChange` to `AccessibleField`, I'd
probably expect it to be forwarded to the child, not set on the wrapper;
and the difference would be particularly jarring for custom components
that can't bubble events.

Rather than build some kind of complicated white- and/or blacklist, I
decided to compromise: I forward all the props (except the accessibility
ones already going to the child), but I left the Typescript and
PropTypes definitions the same. This fulfills the basic ticket objective
of allowing `data-*` attributes (since Typescript ignores those) while
discouraging use of other props.

In order to exclude style-only props, I also added a new helper for
working with styles. The way it works with Typescript is weird, but it
omits all the style props before they get to the inner function.