feat(cactus-web): add prop-forwarding to BrandBar
CACTUS-597

Mostly it was already there. I just had to make one tweak to a component
that can actually be broken by prop forwarding (and why the heck do they
even allow that?), but for the most part I went with the same compromise
as for AccessibleField: forward the props, but don't advertise in types.