feat(cactus-web): add data-prop-forwarding to Tooltip
CACTUS-597

I'm a lot less confident in my choice of which element to forward props
to, which is why I only did the data props. Depending on what you're
testing, you might want the icon or the popup instead; I picked the
hidden label because unlike the popup it's always in the document, and
it already has the `id` prop.