fix(cactus-web): make focus/blur behave more like builtin elements
CACTUS-612

Doing the `toggle` in the `requestAnimationFrame` allows you to do
things like use devtools while keeping the dialog open. However, it
occurred to me that the reason that's necessary is that the
activeElement emits a blur event in those situations even though it's
still the activeElement, so really, custom inputs should do the same.
Thus, it'll fire blur/focus events when the window loses focus but only
toggle the dialog when the activeElement changes, bringing its behavior
more in line with how actual DOM elements work.