fix(cactus-web): fix code for overriding CSS class
I'm adding this kind of thing to make it easier to override styles on
"inner" components, where the user won't have access to the actual
component object to get the class name. Probably won't ever be used, to
be honest, but it tickles my "usability" bone. (Which is like a funny
bone, but less humerus...)