fix(cactus-theme): change shadow fallback to outline
CACTUS-664

I was working on Calendar when I noticed that even when I set the
`border-radius` to the same values for the outer and inner divs, they
didn't line up. Turns out `border-radius` is the outer radius, which
means the inner radius changes when you add a border. As such, outline
is a closer analogue of box-shadow, because it does not affect the
dimensions of the container in any way.