fix(cactus-web): fix layout component sorting
CACTUS-699

The problem was that JavaScript's sort algorithm assumes a homogenous
array (which is a good assumption, normally), so it doesn't have to
compare every item to every other item: if the order went grid, fixed,
grid, then both grid items might be considered "equal to" the fixed item
and even if the first grid item is actually "greater than" the second,
they'll be left in their original position.

My solution is to just insert new items into the correct sorted order,
using an algorithm that only compares items of the same type. Thus fixed
items won't affect the ordering of grid items and vice-versa.