fix(cactus-web): allow '' or null to clear a DateInput
This works by introducing a sentinel value (NaN) to represent partial or invalid dates in the onChange event. When the sentinel value is passed back to the component it is ignored, allowing users to make incremental changes while still keeping any external code (e.g. forms) up-to-date with the input's value & validity.

BREAKING CHANGE: When the current value is partial or invalid, the onChange event.target.value is NaN; previously it was either the empty string or `new Date(NaN)`.

CACTUS-719