fix(cactus-web): allow margin overrides on Accordion.body
CACTUS-720

Part of it was already possible, the rest was switching the default
padding to default margins so they could be overridden.

Also, to make it easier to apply the styles where I wanted them I ripped
out the ridiculously complicated animation code and replaced it with a
`react-transition-group` Transtion.