refactor(cactus-web): improve the Card styles/story
The Card story looks a little weird on large screens, so we wanted to make it look better by default. While I was at it, I figured I'd fix the Card component itself with the latest styling helpers. Since it already had width, I added maxWidth to use in the story as a way of making it narrower.

CACTUS-799