test(cactus-web): add tests to illustrate RadioButton ref support
Mostly for Typescript purposes. As per usual I also refactored the
styles a bit. The biggest difference here is that I made it so `id` is
no longer required; I assume it was originally required to connect the
label to the hidden input, but now that the label wraps the input there
doesn't need to be an explicit connection.

CACTUS-782