test(cactus-web): add tests to illustrate CheckBox ref support
Mostly for Typescript purposes. As per usual I also refactored the
styles a bit. Like with RadioButton, I made `id` optional. This one also
had an oddity in that `name` was required on CheckBoxField (but not
CheckBox). It makes some sense to require `name` on radios, since
they're basically non-functional without a name, but CheckBox doesn't
need a `name` to work so I made that optional as well.

CACTUS-782