refactor(cactus-form): utilize `patch-package` to fix react-final-form
This is now framed as an optional (but recommended) step, because
technically everything in `cactus-form` does work just fine without it,
it's only some interactions with specific `cactus-web` components that
suffer. Even so, the amount of extra setup is no greater than
patch-package itselt, so I think it's about as good as we can get.

CACTUS-727