chore: upgrade `@repay/scripts`
So I managed to narrow down the library causing our tests to hang to
`core-js >= 3.16.0`. I attempted to scope only the version used by
`@repay/scripts` in the resolution (which should be possible according
to the docs), but it didn't work. Gatsby is the only other package that
requests a version higher than 3.16 and it still appears to work, so
there shouldn't be any issue with pinning the version (for now).
Still eagerly awaiting the day we stop supporting IE.

CACTUS-880