feat(cactus-web): pagination dynamically changes button count by width
This changes how upper/lower breaks are calculated. The change to the lower break is minor, and
mostly due to the fact that now upper break is calculated first. The change to upper break is
significant in that it now replaces the last page number, instead of replacing UP TO the last page.
This means you can no longer tell what the last page is at a glance, but I felt trying to mimic the
old behavior when the number of pages is dynamic would be far too complicated.

CACTUS-780