fix: fix visual tests, rename Docker image tags
As far as I can tell, the visual tests were still working in GitHub because the Docker container didn't mess with yarn at all: instead it just set up the container OS and used the entire repo as a volume (I think; it's not explained clearly in the GitHub Actions docs). So now we do the same.

While I was at it I reduced the duplication by splitting it into two images, and renamed the images to `cactus/tests` so they're less likely to conflict with any official or custom local image tags.

CACTUS-915