feat(cactus-web): add flex layout options to Footer component
Default behavior stays the same, uses the `flexFlow` prop to opt-in to
the flex behavior. A fair amount of the changes are just cleaning up the
styles to use the style helpers. Also didn't need the separate sections
for `link`/`lightLink`, when the `variant` section could just override
them instead.

No idea what's wrong with the stupid docgen, but it couldn't find the
`Footer.Logo` component, and it was attaching the logo style props to
the base `Footer` component. As such I rewrote the props section.

CACTUS-925