refactor(cactus-i18n): fix types & behavior for React 18
The toggle strategy for forcing a re-render is flawed in React 18,
because it can group several state changes into a single render: if two
such changes are grouped, then the `flipFlop` variable will end with the
same value as it started and the context won't be updated. (It's a good
thing the standard app has an even number of sections or I wouldn't have
noticed the bug.)

I was going to fix it by just changing from toggling a boolean to
incrementing an integer, but then I came across a reference to inline
dispatch/setState which was a much more straightforward solution than
the weird auxiliary ref change tracking I was doing before.

CACTUS-926