refactor(cactus-web): fix types for React 18
Pretty much all just adding or fixing the type of `children` props,
although Calendar had an interesting bug that wasn't caught in React 17
because `ReactNode` included the `{}` type, which basically matches
everything so it wouldn't catch incorrect types.

I mostly fixed the effect hooks in separate commits, but there was one
hook that was theoretically bad but shouldn't cause problems in
practice, so I threw in a comment on why I'm not changing that one.

CACTUS-926