fix(cactus-web): make Calendar.Grid a little more React 18 proof
This references the offscreen API, which supposedly "unmounts"
components without destroying the state. Unfortunately I haven't been
able to find anything definitive in the docs regarding what happens to
refs: it runs unmount effects which would imply setting the refs to
null, but on the other hand refs aren't exactly effects and if the
purpose is to make rendering faster, it would make sense to keep the DOM
elements around but detached from the document tree. When we actually
start supporting React 18 we'll need to figure out how to test that, to
see if we need an auxiliary `isMounted` ref controlled only by effects.

CACTUS-926