fix(cactus-web): change render trigger to use unique values
Because of how React 18 more intelligently groups renders, several state
changes may end up queued into a single render. With the toggle reducer,
that means the state value might end up the same as it was initially. We
previously had a component that did something similar but used the value
in a dependency array, leading to the effect hook not firing even though
the value "changed" (because it changed back before the actual render).

Note that in this case that doesn't actually matter because we ignore
the value and the re-render is still triggered; I'm making this change
so that people looking at the code will have a better example of how to
implement this kind of pattern.

CACTUS-926