fix(cactus-icons): don't forward style props to SVG element
I could have just stuck this logic in the template, but frankly there
was already way more duplicated code than we should have in there.
Instead I decided to extract all the common styles & logic into an
"abstract" base icon. Then I thought I might as well export it so that
other projects can easily add their own icons if they need to.

I removed the `propTypes` because it was inconsistent: `iconSizes` has
basically the same constraints as all the other styled props, but it was
the only one we explicitly had types for. I assume that originally it's
because we only "officially" supported the size strings ('tiny', etc.);
then when we changed that to support any number or string we expanded
the prop type even though it was no longer very informative.

I also added `forwardRef`, which is completely unrelated but I figured I
might as well since I was already rewriting the template.

CACTUS-931