feat(cactus-web): improve DataGrid Column interface
The main change is that I combined Column and DataColumn into a single
component. To make that a little easier I moved most of the column logic
into a separate file, and then rewrote the rendering logic in
DataGridTable. I doubt we'd ever see enough columns where the performance
difference between a Map and an Array would be significant, so I got rid
of the Map and just kept the whole thing as a sorted array.

With the table rendering, that was really too complex to keep inline the
way it was. Although it's a bit simpler now that we have only one column
type, I still thought it best to split the header & body cell logic into
separate functions, and I tried to simplify the sorting logic a bit.

I was originally going to try to get rid of the column state entirely
and just render everything in a single pass by relying on the fact that
React seems to render depth first: I'd pass an empty array to the
context which would be filled in the render func of the Columns, then it
would move on to render the Table with the now-filled array of columns.
In the end, though, I felt there was too much that could go wrong; and in
particular React 18 broke some of my earlier attempts at improved
efficiency, so I figured I should probably stick to the contract and just
avoid all side effects in render functions.

- Added `order` prop, defaults to column render order.
- Added `headerProps` prop, to set props on just the header cells.
- Fixed `isCardView` variable to take `variant` into account.
- Changed DataGridType, using the actual component types instead of generic ComponentType.
- Removed/simplified some styles:
  - `IconWrapper` only added `flex-shrink: 1`, which is a default style.
  - `TextWrapper` added `flex-grow: 1`, but since the flex container itself is doesn't grow, that does nothing as far as I could tell (even on IE).
  - `.flip-chevron` class replaced by simply rendering `NavigationChevronUp` instead of transforming the Down icon.
  - `.has-content` class replaced using the `:empty` pseudo-class.
  - Defined the TopSection's `spacing` as CSS, instead of cloning & adding a style prop like it was before.

I really wanted to get rid of that `topSectionRendered` weirdness, but I
felt that was just too far out of scope (as opposed to all the other
changes I made to `TopSection` ;).

CACTUS-964