fix(cactus-icons): rename CactusIcon to AbstractIcon
We have a babel plugin that automatically sets the displayName on styled
components; while `AbstractIcon` is a better name for the exported
component, the CSS class makes more sense as `CactusIcon` since that
appears on all the derived icons as well. As such I had to do a little
renaming workaround.

CACTUS-931