test(cactus-web): update storyshots
Not sure why the Breadcrumb change suddenly happened, but this seems
more correct: if the list CAN fit in the parent, why shouldn't it do so?

The other change is due to the change from `.has-content` to `:empty`;
they should have been equal, but I guess `.has-content` wasn't
calculated correctly or something?

CACTUS-964