docs(cactus-web): update Grid docs with CSS grid support
I am so tired of how flaky the automatic props table is. I don't think
I've had a single time where it "just worked", and frankly, I no longer
think it is in any way worth it. Maintining the docs this way is a
little more work, but much more flexible: we can show more relevant
information, and we don't have to try to put it all in one-size-fits-all
tables.

CACTUS-972