fix(cactus-web): implement default styles statically
Using `defaultProps` when those map to inline styles makes style
overrides more likely to break. Technically that's already acknowledged
as a breaking change here, but no need to make it worse than it has to be.

Spinner has a similar issue, but it's basically an icon so that change
will need to be made in conjunction with a similar update to cactus-icons.

CACTUS-1069