feat(cactus-web): change SplitButton interface
BREAKING CHANGE: (1) SplitButton no longer has separate props for defining the main action; instead all actions are defined as children, and the main action is the either the first action with the `main` prop, or the first action in render order if none have the `main` prop. (2) `icon` prop was removed; icons will be passed as part of the action's children. (3) `onSelect` changed to `onClick`; Typescript users should not expect this change to automatically break the types, because `SplitButton.Action` supports all the normal DOM events, and there is still an `onSelect` handler, but as expected it now handles `select` events rather than `click` events.

CACTUS-767