refactor(cactus-web): update Modal to inline styles
This one was a bit tricky because of the extra percent-to-fixed conversion.
It was also the last usage of the old styled helpers, so they have been removed.

Also had to change Select's options prop to readonly; mutable array can be
assigned to readonly, but not vice versa, so this shouldn't break anything.

CACTUS-1093